feat: fix tabs navigation bugs and improve tabs organization for 2 profiles
