+++ src/Screens/CustomerOrderDetail/index.tsx
@@ -124,7 +124,7 @@ export function CustomerOrderDetail({ orderId }: OrderDetailProps) {
             order_status: 3,
           };
           //sendDeliveredOrder(orderToQueue);
-          router.push("/(deliver)/deliveries");
+          router.push("/(deliver)/(tabs)/deliveries");
         },
       },
     ]);
